com.unity.muse.common@0.2.0 ## [0.2.0] - 2023-09-20
### Changed

- Improve in-painting experience.
- Artifacts cannot be unselected in refinement.

### Fixed

- Fix delete not appearing in context menu when there was an error with the generation.
- Fix title of Muse window reseting when reloading the window.
- Fix error when working with a large number of selected assets.
- Fix artifact favorite icon not showing after changing the thumbnail.

+++ Runtime/Artifacts/ArtifactCache/LiteDB/LiteDbArtifactCache.cs
@@ -97,7 +97,7 @@ namespace Unity.Muse.Common.Cache
                 {
                     case "png":
                     {
-                        var texture = new Texture2D(2, 2);
+                        var texture = TextureUtils.Create();
                         texture.LoadImage(dbObject.RawData);
 
                         return texture;
@@ -138,7 +138,14 @@ namespace Unity.Muse.Common.Cache
             {
                 foreach (var artifact in artifacts)
                 {
+                    if(artifact == null)
+                        continue;
+
                     var dbArtifact = collection.FindOne(x => x.Guid == artifact.Guid);
+
+                    if(dbArtifact == null)
+                        continue;
+
                     collection.Delete(dbArtifact.Id);
                 }
